+++ CHANGELOG.md
@@ -1,3 +1,12 @@
-## 0.1.0 (Unreleased)
+## 0.2.0 - 2026-02-06
+### Added
+* `open_tracking_enabled` and `click_tracking_enabled` are now configurable properties
+* When creating, updating or refreshing a domain, a check request is sent in order to
+  get up-to-date verification information.
 
-FEATURES:
+### Fixed
+* Non-20x HTTP Status codes now reliably lead to errors
+
+## 0.1.0 - 2026-01-31
+### Added
+* The whole provider